Move override_source field to codec configuration.
Let codecs extend the AbstractCodec.Config so every codec gets the
default override_source field.
