+++ graylog2-inputs/src/main/java/org/graylog2/inputs/codecs/RadioMessageCodec.java
@@ -94,12 +94,7 @@ public class RadioMessageCodec extends AbstractCodec {
     }
 
     @ConfigClass
-    public static class Config implements AbstractCodec.Config {
-        @Override
-        public ConfigurationRequest getRequestedConfiguration() {
-            return new ConfigurationRequest();
-        }
-
+    public static class Config extends AbstractCodec.Config {
         @Override
         public void overrideDefaultValues(@Nonnull ConfigurationRequest cr) {
