+++ graylog2-plugin-interfaces/src/main/java/org/graylog2/plugin/inputs/codecs/AbstractCodec.java
@@ -23,6 +23,9 @@
 package org.graylog2.plugin.inputs.codecs;
 
 import org.graylog2.plugin.configuration.Configuration;
+import org.graylog2.plugin.configuration.ConfigurationRequest;
+import org.graylog2.plugin.configuration.fields.ConfigurationField;
+import org.graylog2.plugin.configuration.fields.TextField;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -57,5 +60,28 @@ public abstract class AbstractCodec implements Codec {
             }
         }
         return name;
+    }
+
+    public abstract static class Config implements Codec.Config {
+        @Override
+        public ConfigurationRequest getRequestedConfiguration() {
+            final ConfigurationRequest configurationRequest = new ConfigurationRequest();
+
+            // TODO implement universal override (in raw message maybe?)
+            configurationRequest.addField(new TextField(
+                    CK_OVERRIDE_SOURCE,
+                    "Override source",
+                    null,
+                    "The source is a hostname derived from the received packet by default. Set this if you want to override " +
+                            "it with a custom string.",
+                    ConfigurationField.Optional.OPTIONAL
+            ));
+
+            return configurationRequest;
+        }
+
+        @Override
+        public void overrideDefaultValues(@Nonnull ConfigurationRequest cr) {
+        }
     }
 }
